feat:関数の切り分け
